+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ChatApp.kt
@@ -5,7 +5,11 @@ import com.getstream.sdk.chat.utils.DateFormatter
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.logger.ChatLogLevel
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.offline.ChatDomain
+import io.getstream.chat.android.client.notifications.handler.NotificationConfig
+import io.getstream.chat.android.client.notifications.handler.NotificationHandler
+import io.getstream.chat.android.client.notifications.handler.NotificationHandlerFactory
+import io.getstream.chat.android.livedata.ChatDomain
+import io.getstream.chat.android.pushprovider.firebase.FirebasePushDeviceGenerator
 
 class ChatApp : Application() {
 
@@ -18,24 +22,48 @@ class ChatApp : Application() {
         super.onCreate()
         dateFormatter = DateFormatter.from(this)
 
+        setupStreamSdk()
+        connectUser()
+    }
+
+    private fun setupStreamSdk() {
         val client = ChatClient.Builder("qx5us2v6xvmh", applicationContext)
             .logLevel(ChatLogLevel.ALL)
+            .notifications(createNotificationConfig(), createNotificationHandler())
             .build()
         ChatDomain.Builder(client, applicationContext)
             .userPresenceEnabled()
             .build()
+    }
 
-        val user = User(
-            id = "jc",
-            extraData = mutableMapOf(
-                "name" to "Jc Miñarro",
-                "image" to "https://ca.slack-edge.com/T02RM6X6B-U011KEXDPB2-891dbb8df64f-128",
+    private fun connectUser() {
+        ChatClient.instance().connectUser(
+            user = User(
+                id = "filip",
+                extraData = mutableMapOf(
+                    "name" to "Filip Babić",
+                    "image" to "https://ca.slack-edge.com/T02RM6X6B-U022AFX9D2S-f7bcb3d56180-128",
+                ),
             ),
-        )
-
-        client.connectUser(
-            user = user,
-            token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiamMifQ.2_5Hae3LKjVSfA0gQxXlZn54Bq6xDlhjPx2J7azUNB4"
+            token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiZmlsaXAifQ.WKqTjU6fHHjtFej-sUqS2ml3Rvdqn4Ptrf7jfKqzFgU"
         ).enqueue()
+    }
+
+    private fun createNotificationConfig(): NotificationConfig {
+        return NotificationConfig(
+            pushDeviceGenerators = listOf(FirebasePushDeviceGenerator())
+        )
+    }
+
+    private fun createNotificationHandler(): NotificationHandler {
+        return NotificationHandlerFactory.createNotificationHandler(
+            context = this,
+            newMessageIntent = { _: String, channelType: String, channelId: String ->
+                PushHandlerActivity.getIntent(
+                    context = this,
+                    channelId = "$channelType:$channelId"
+                )
+            }
+        )
     }
 }